Time: 11 ms (22.4%), Space: 9.2 MB (56.82%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.c
@@ -19,7 +19,7 @@ int minEatingSpeed(int* piles, int pilesSize, int h) {
       long long needed_hours = 0;
       
       for(int i=0; i < pilesSize; i++){
-        needed_hours = needed_hours + ((long long)(piles[i] + mid - 1) / mid);
+        needed_hours = needed_hours + ((piles[i] + mid - 1) / mid);
       }
       
       if(needed_hours <= h){
